:sparkles: Tratando exceções

+++ src/storage/group/groupCreate.ts
@@ -1,11 +1,19 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { GROUP_COLLECTION } from "@storage/storageConfig";
+import { AppError } from "@utils/AppError";
 
 import { groupsGetAll } from "./groupsGetAll";
 
 export async function groupCreate(newGroup: string) {
   try {
     const storedGroups = await groupsGetAll();
+
+    const groupExists = storedGroups.includes(newGroup);
+
+    if (groupExists) {
+      throw new AppError("Já existe um grupo com esse nome.");
+    }
+
     await AsyncStorage.setItem(
       GROUP_COLLECTION,
       JSON.stringify([...storedGroups, newGroup])
